BC-2770 Fix type and migration

+++ apps/server/src/modules/files-storage/entity/filerecord.entity.ts
@@ -87,7 +87,7 @@ export class FileRecord extends BaseEntity {
 	@Property({ nullable: true })
 	deletedSince?: Date;
 
-	@Property({ type: types.bigint })
+	@Property({ type: types.integer })
 	size: number;
 
 	@Property()
